@@ -1,26 +1,18 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
+export default {
   content: [
     "./index.html",
     "./src/**/*.{vue,js,ts,jsx,tsx}",
     "./node_modules/tw-elements/dist/js/**/*.js",
-    "./node_modules/flowbite/**/*.js"
+    "./node_modules/flowbite/**/*.js",
   ],
   theme: {
     extend: {},
-    listStyleType: {
-      none: 'none',
-      disc: 'disc',
-      decimal: 'decimal',
-      square: 'square',
-      roman: 'upper-roman',
-    }
   },
-  darkMode: "class",
   plugins: [
-    require("tw-elements/dist/plugin.cjs"),
     require('flowbite/plugin'),
-    require('@tailwindcss/typography'),
-  ]
+    require("tw-elements/dist/plugin.cjs")
+  ],
+  darkMode: "class"
 }
 
