@@ -1,8 +1,19 @@
-import { createApp } from 'vue'
-import App from './App.vue'
-import router from './router'
-import store from './store'
+import { createApp, markRaw } from 'vue';
+import { createPinia } from 'pinia';
 
-import './assets/css/style.css'
+import App from './App.vue';
+import router from './router';
+import './plugins/axios.js';
+import './plugins/vee-validate.js';
 
-createApp(App).use(store).use(router).mount('#app')
+import './assets/main.css';
+
+const pinia = createPinia();
+pinia.use(({ store }) => {
+  store.router = markRaw(router);
+});
+
+const app = createApp(App);
+app.use(pinia);
+app.use(router);
+app.mount('#app');
